misc things
did alot of misc things that should hopefuly decrease bundle size
